update dependencies and moved navigation to a separate package
