+++ app/src/main/java/com/hadiyarajesh/composetemplate/ui/App.kt
@@ -5,7 +5,7 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.rememberNavController
-import com.hadiyarajesh.composetemplate.ui.navigation.AppNavigation
+import com.hadiyarajesh.composetemplate.navigation.AppNavigation
 import com.hadiyarajesh.composetemplate.ui.theme.AppTheme
 
 @Composable
