+++ app/src/main/java/com/hadiyarajesh/composetemplate/ui/home/HomeScreen.kt
@@ -18,10 +18,10 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import com.hadiyarajesh.composetemplate.R
 import com.hadiyarajesh.composetemplate.data.entity.Message
+import com.hadiyarajesh.composetemplate.navigation.TopLevelDestination
 import com.hadiyarajesh.composetemplate.ui.components.ErrorItem
 import com.hadiyarajesh.composetemplate.ui.components.LoadingIndicator
 import com.hadiyarajesh.composetemplate.ui.components.VerticalSpacer
-import com.hadiyarajesh.composetemplate.ui.navigation.TopLevelDestination
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
