+++ app/src/main/java/com/hadiyarajesh/composetemplate/ui/home/HomeScreenUiState.kt
@@ -2,6 +2,9 @@ package com.hadiyarajesh.composetemplate.ui.home
 
 import com.hadiyarajesh.composetemplate.data.entity.Message
 
+/**
+ * Sealed class to represent UI states in [HomeScreen]
+ */
 sealed interface HomeScreenUiState {
     data object Initial : HomeScreenUiState
     data object Loading : HomeScreenUiState
