+++ app/src/main/java/com/hadiyarajesh/composetemplate/utility/Constants.kt
@@ -1,6 +1,6 @@
 package com.hadiyarajesh.composetemplate.utility
 
 object Constants {
-    const val API_BASE_URL = "https://api.github.com/"
+    const val API_BASE_URL = "API_BASE_URL"
     const val SOURCE = "source"
 }
