+++ gradle/libs.versions.toml
@@ -2,7 +2,7 @@
 agp = "8.1.2"
 kotlin = "1.9.0"
 hilt = "2.48"
-hilt-navigation = "1.0.0"
+hilt-navigation = "1.1.0"
 ksp = "1.9.0-1.0.13"
 core-ktx = "1.12.0"
 junit = "4.13.2"
@@ -10,10 +10,10 @@ androidx-test-ext-junit = "1.1.5"
 espresso-core = "3.5.1"
 lifecycle-runtime-ktx = "2.6.2"
 activity-compose = "1.8.0"
-compose-bom = "2023.10.00"
+compose-bom = "2023.10.01"
 compose-compiler = "1.5.1"
-navigation-compose = "2.7.4"
-room = "2.5.2"
+navigation-compose = "2.7.5"
+room = "2.6.0"
 retrofit = "2.9.0"
 okhttp-logging-interceptor = "4.10.0"
 moshi = "1.14.0"
